Remove validation webhook
This was optional and disable by default. We are introducing cel validation and there's no known use case for this webhook today. We don't want to maintain this base code nor support validation/default webhooks operational burden
